[DEV-2313] Improved verified name check
